requirements.txt
openai-whisper
jiwer
pyaudio
colorama
